fixed the only thing i don't like about the app!
Previously:
when you search for a song and pick it, the app creates a playlist with just that song.

Now:
when you pick a searched for song, the app skips to that song on the current playlist. And if the song isn't on the current playlist then it should start the all songs playlist on that particular song but that's not done yet